+++ lib/src/analyzers/lint_analyzer/rules/rules_list/no_equal_arguments/config_parser.dart
@@ -1,0 +1,11 @@
+part of 'no_equal_arguments.dart';
+
+class _ConfigParser {
+  static const _ignoredParametersConfig = 'ignored-parameters';
+
+  static Iterable<String> parseIgnoredParameters(Map<String, Object> config) =>
+      config.containsKey(_ignoredParametersConfig) &&
+              config[_ignoredParametersConfig] is Iterable
+          ? List<String>.from(config[_ignoredParametersConfig] as Iterable)
+          : <String>[];
+}