+++ test/analyzers/lint_analyzer/rules/rules_list/no_equal_arguments/examples/named_parameters_example.dart
@@ -1,0 +1,10 @@
+String getUserImage({String firstName, String lastName}) {
+  return '/test_url/' + firstName ?? '' + lastName ?? '';
+}
+
+const firstName = 'name';
+
+final image = getUserImage(
+  firstName: firstName,
+  lastName: firstName, // LINT
+);